% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ch_token.R
\name{get_ch_token}
\alias{get_ch_token}
\title{get_ch_token Function:}
\usage{
get_ch_token(uri, resource, client_id, client_secret)
}
\arguments{
\item{uri}{Oauth2 server address}

\item{resource}{Oauth2 server resource name}

\item{client_id}{is the Cloud Historian AppID}

\item{client_secret}{is the Cloud Historian AppID Secret}
}
\value{
a security token
}
\description{
A function to connect to an oauth2 server and return a security token
}
