% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCHData.R
\name{getCHData}
\alias{getCHData}
\title{getCHData Function:}
\usage{
getCHData(uri, token, start_time, end_time, tag_names, down_sample,
  sys_guid)
}
\arguments{
\item{uri}{Cloud Historian Server Address}

\item{token}{Oauth2 security token. Use getCHToken to acquire a token}

\item{start_time}{The start time to be used for data retrieval}

\item{end_time}{The start time to be used for data retrieval}

\item{tag_names}{A vector of tagnems to be used for data retrieval}

\item{down_sample}{The the downsampling function to be used in the data retrieval
Valid responses:
Downsample to 1 min average: "1m-avg",
Downsample to 60 second average: "60s-avg",
Downsample to 1 hour average: "1h-avg",
Downsample to avg of all results: "0all-avg",
Downsample to max of all results: "0all-max",
return raw data: ""}
}
\value{
An object with the request information and the formated data in the $data attribute.
}
\description{
A function to connect to Honewell Sentience Cloud Historian Server and return the requested timeseries data
}
